change barcode to long (db bigint type)

+++ src/dataAccess/IProductDAO.java
@@ -8,10 +8,10 @@ import java.util.ArrayList;
 
 public interface IProductDAO {
     int create(Product product) throws InsertionFailedException, DAORetrievalFailedException;
-    int delete(int barcode) throws DeleteFailedException, DAORetrievalFailedException;
+    int delete(Long barcode) throws DeleteFailedException, DAORetrievalFailedException;
     int update(Product product) throws UpdateFailedException, DAORetrievalFailedException;
 
-    Product findByBarcode(int barcode) throws NotFoundException, DAORetrievalFailedException;
+    Product findByBarcode(Long barcode) throws NotFoundException, DAORetrievalFailedException;
     ArrayList<Product> findByName(String name) throws NotFoundException, DAORetrievalFailedException;
     ArrayList<Product> findAll() throws DAORetrievalFailedException;
 
@@ -20,6 +20,6 @@ public interface IProductDAO {
     ArrayList<Category> getAllCategories() throws DAORetrievalFailedException;
     ArrayList<Vat> getAllVatTypes() throws DAORetrievalFailedException;
 
-    int currentStock(int barcode) throws NotFoundException, DAORetrievalFailedException;
+    int currentStock(Long barcode) throws NotFoundException, DAORetrievalFailedException;
     ArrayList<Integer> outOfStock() throws NotFoundException, DAORetrievalFailedException;
 }
